requirements.txt
flask
flask-cors
supabase
openai
gunicorn
python-dotenv
requests
pypdf
pdfminer.six
tiktoken
flask
flask-cors
flask-login
gunicorn
openai
pdfminer.six
pypdf
python-dotenv
requests
supabase
tiktoken
werkzeug
flask
flask-cors
gunicorn
openai
pdfminer.six
pypdf
python-dotenv
requests
supabase
tiktoken
boto3
